Split news HTML rendering
